+++ src/css/App.css
@@ -319,7 +319,7 @@ div.game-grid-container div.game-grid-item button span.btn-placeholder {
 }
 
 div.game-grid-container div.game-grid-item:hover {
-  background: rgba(255, 255, 255, 0.05);
+  background: rgba(255, 255, 255, 0.07);
 }
 
 div.game-grid-container div.game-grid-item:hover span.btn-placeholder {
